requirements.txt
google-cloud-aiplatform[agent_engines,adk]==1.126.1
google-adk[eval, a2a]==1.18.0
cloudpickle
google-genai
google-cloud-discoveryengine
pydantic
python-dotenv
google-auth-oauthlib
typer
fastapi
fastmcp